binding/c: autogen op api
