+++ src/binding/c/op_api.txt
@@ -1,0 +1,25 @@
+# vim: set ft=c:
+
+MPI_Op_create:
+    .desc: Creates a user-defined combination function handle
+    .extra: collops
+    .seealso: MPI_Op_free
+/*
+    Notes on the user function:
+    The calling list for the user function type is
+    .vb
+        typedef void (MPI_User_function) (void * a,
+                      void * b, int * len, MPI_Datatype *);
+    .ve
+    where the operation is 'b[i] = a[i] op b[i]', for 'i=0,...,len-1'.  A pointer
+    to the datatype given to the MPI collective computation routine (i.e.,
+    'MPI_Reduce', 'MPI_Allreduce', 'MPI_Scan', or 'MPI_Reduce_scatter') is also
+    passed to the user-specified routine.
+*/
+
+MPI_Op_free:
+    .desc: Frees a user-defined combination function handle
+
+MPI_Op_commutative:
+    .desc: Queries an MPI reduction operation for its commutativity
+    .extra: NULL
