+++ src/mpi/coll/op/Makefile.mk
@@ -3,15 +3,8 @@
 ##     See COPYRIGHT in top-level directory
 ##
 
-# mpi_sources includes only the routines that are MPI function entry points
-# The code for the MPI operations (e.g., MPI_SUM) is not included in
-# mpi_sources
-mpi_sources +=                     \
-    src/mpi/coll/op/op_create.c       \
-    src/mpi/coll/op/op_free.c         \
-    src/mpi/coll/op/op_commutative.c
-
 mpi_core_sources += \
+    src/mpi/coll/op/op_impl.c        \
     src/mpi/coll/op/oputil.c         \
     src/mpi/coll/op/opsum.c          \
     src/mpi/coll/op/opmax.c          \
